--- examples/ts/cli/.env
@@ -1,3 +1,0 @@
-INWORLD_KEY=w8NUknyW0UaPWeH0z6YzGH7hGCxygbHW
-INWORLD_SECRET=tSPcQb45dJQPEfKyskxsL3scDzMb4ppgnlMZH2y6Bic3OsBr5U5N48PQ5MOIVl9I
-INWORLD_SCENE=workspaces/dunes_of_aurora_campaign/scenes/dune_cliffs
