+++ examples/ts/cli/.env-sample
@@ -1,0 +1,5 @@
+# DO NOT EDIT THIS FILE DIRECTLY
+# Copy this file to .env and enter your Inworld API Key, Secret and the Scene ID
+INWORLD_KEY=
+INWORLD_SECRET=
+INWORLD_SCENE=
